pyproject.toml
[project]
name = "timesheeterpython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.3",
    "flask-wtf>=1.2.2",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "pandas>=2.3.3",
    "pypdf>=6.7.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "tabulate>=0.9.0",
    "wtforms>=3.2.1",
    "xhtml2pdf>=0.2.17",
]
